% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodbstreams_operations.R
\name{get_shard_iterator}
\alias{get_shard_iterator}
\title{Returns a shard iterator}
\usage{
get_shard_iterator(StreamArn, ShardId, ShardIteratorType,
  SequenceNumber = NULL)
}
\arguments{
\item{StreamArn}{The Amazon Resource Name (ARN) for the stream.}

\item{ShardId}{The identifier of the shard. The iterator will be returned for this shard ID.}

\item{ShardIteratorType}{Determines how the shard iterator is used to start reading stream records from the shard:
\itemize{
\item \code{AT_SEQUENCE_NUMBER} - Start reading exactly from the position denoted by a specific sequence number.
\item \code{AFTER_SEQUENCE_NUMBER} - Start reading right after the position denoted by a specific sequence number.
\item \code{TRIM_HORIZON} - Start reading at the last (untrimmed) stream record, which is the oldest record in the shard. In DynamoDB Streams, there is a 24 hour limit on data retention. Stream records whose age exceeds this limit are subject to removal (trimming) from the stream.
\item \code{LATEST} - Start reading just after the most recent stream record in the shard, so that you always read the most recent data in the shard.
}}

\item{SequenceNumber}{The sequence number of a stream record in the shard from which to start reading.}
}
\description{
Returns a shard iterator. A shard iterator provides information about how to retrieve the stream records from within a shard. Use the shard iterator in a subsequent \code{GetRecords} request to read the stream records from the shard.
}
\details{
A shard iterator expires 15 minutes after it is returned to the requester.
}
\examples{
# The following example returns a shard iterator for the provided stream
# ARN and shard ID.
get_shard_iterator(
  ShardId = "00000001414576573621-f55eea83",
  ShardIteratorType = "TRIM_HORIZON",
  StreamArn = "arn:aws:dynamodb:us-west-2:111122223333:table/Forum/stream/2015-05-20T20:51:10.252"
)

}
