% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodbstreams_package.R
\docType{package}
\name{paws.dynamodbstreams-package}
\alias{paws.dynamodbstreams}
\alias{paws.dynamodbstreams-package}
\title{paws.dynamodbstreams: Amazon DynamoDB Streams}
\description{
Amazon DynamoDB Streams provides API actions for
    accessing streams and processing stream records. To learn more about
    application development with Streams, see [Capturing Table Activity
    with DynamoDB
    Streams](http://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Streams.html)
    in the Amazon DynamoDB Developer Guide.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
